+++ packages/gatsby/src/redux/types.ts
@@ -38,13 +38,13 @@ export interface IGatsbyPage {
 }
 
 export interface IGatsbyConfig {
-  plugins?: {
+  plugins?: Array<{
     // This is the name of the plugin like `gatsby-plugin-manifest
     resolve: string
     options: {
       [key: string]: unknown
     }
-  }[]
+  }>
   siteMetadata?: {
     title?: string
     author?: string
@@ -64,7 +64,7 @@ export interface IGatsbyConfig {
 export interface IGatsbyNode {
   id: Identifier
   parent: Identifier
-  children: Identifier[]
+  children: Array<Identifier>
   internal: {
     type: string
     counter: number
@@ -76,7 +76,7 @@ export interface IGatsbyNode {
   }
   __gatsby_resolved: any // TODO
   [key: string]: unknown
-  fields: string[]
+  fields: Array<string>
 }
 
 export interface IGatsbyPlugin {
@@ -87,7 +87,7 @@ export interface IGatsbyPlugin {
 }
 
 export interface IGatsbyPluginContext {
-  [key: string]: (...args: any[]) => any
+  [key: string]: (...args: Array<any>) => any
 }
 
 export interface IGatsbyStaticQueryComponents {
@@ -121,8 +121,8 @@ export interface IPlugin {
 }
 
 export interface IBabelStage {
-  plugins: IPlugin[]
-  presets: IPlugin[]
+  plugins: Array<IPlugin>
+  presets: Array<IPlugin>
   options: {
     cacheDirectory: boolean
     sourceType: string
@@ -137,7 +137,7 @@ type BabelStageKeys =
   | "build-javascript"
 
 export interface IStateProgram extends IProgram {
-  extensions: string[]
+  extensions: Array<string>
 }
 
 export interface IGatsbyState {
@@ -147,7 +147,7 @@ export interface IGatsbyState {
   resolvedNodesCache: Map<string, any> // TODO
   nodesTouched: Set<string>
   lastAction: ActionsUnion
-  flattenedPlugins: {
+  flattenedPlugins: Array<{
     resolve: SystemPath
     id: Identifier
     name: string
@@ -156,7 +156,7 @@ export interface IGatsbyState {
       plugins: []
       [key: string]: unknown
     }
-    nodeAPIs: (
+    nodeAPIs: Array<
       | "onPreBoostrap"
       | "onPostBoostrap"
       | "onCreateWebpackConfig"
@@ -165,16 +165,16 @@ export interface IGatsbyState {
       | "createPagesStatefully"
       | "createPages"
       | "onPostBuild"
-    )[]
-    browserAPIs: (
+    >
+    browserAPIs: Array<
       | "onRouteUpdate"
       | "registerServiceWorker"
       | "onServiceWorkerActive"
       | "onPostPrefetchPathname"
-    )[]
-    ssrAPIs: ("onRenderBody" | "onPreRenderHTML")[]
+    >
+    ssrAPIs: Array<"onRenderBody" | "onPreRenderHTML">
     pluginFilepath: SystemPath
-  }[]
+  }>
   config: IGatsbyConfig
   pages: Map<string, IGatsbyPage>
   schema: GraphQLSchema
@@ -199,15 +199,15 @@ export interface IGatsbyState {
     IGatsbyStaticQueryComponents["id"],
     IGatsbyStaticQueryComponents
   >
-  staticQueriesByTemplate: Map<SystemPath, Identifier[]>
+  staticQueriesByTemplate: Map<SystemPath, Array<Identifier>>
   pendingPageDataWrites: {
     pagePaths: Set<string>
     templatePaths: Set<SystemPath>
   }
   // @deprecated
   jobs: {
-    active: any[] // TODO
-    done: any[] // TODO
+    active: Array<any> // TODO
+    done: Array<any> // TODO
   }
   jobsV2: {
     incomplete: Map<Identifier, IGatsbyIncompleteJobV2>
@@ -215,7 +215,7 @@ export interface IGatsbyState {
   }
   webpack: any // TODO This should be the output from ./utils/webpack.config.js
   webpackCompilationHash: string
-  redirects: IRedirect[]
+  redirects: Array<IRedirect>
   babelrc: {
     stages: {
       [key in BabelStageKeys]: IBabelStage
@@ -231,8 +231,10 @@ export interface IGatsbyState {
       exclude?: { types?: Array<string>; plugins?: Array<string> }
       withFieldTypes?: boolean
     } | null
-    thirdPartySchemas: GraphQLSchema[]
-    types: (string | { typeOrTypeDef: DocumentNode; plugin: IGatsbyPlugin })[]
+    thirdPartySchemas: Array<GraphQLSchema>
+    types: Array<
+      string | { typeOrTypeDef: DocumentNode; plugin: IGatsbyPlugin }
+    >
   }
   themes: any // TODO
   logs: IGatsbyCLIState
@@ -409,7 +411,7 @@ export interface ICreatePageDependencyAction {
 export interface IDeleteComponentDependenciesAction {
   type: "DELETE_COMPONENTS_DEPENDENCIES"
   payload: {
-    paths: string[]
+    paths: Array<string>
   }
 }
 
@@ -496,7 +498,7 @@ export interface ICreateTypes {
   type: `CREATE_TYPES`
   plugin: IGatsbyPlugin
   traceId?: string
-  payload: DocumentNode | DocumentNode[]
+  payload: DocumentNode | Array<DocumentNode>
 }
 
 export interface ICreateFieldExtension {
@@ -586,7 +588,7 @@ export interface ISetStaticQueriesByTemplateAction {
   type: `SET_STATIC_QUERIES_BY_TEMPLATE`
   payload: {
     componentPath: string
-    staticQueryHashes: Identifier[]
+    staticQueryHashes: Array<Identifier>
   }
 }
 
@@ -686,8 +688,8 @@ export interface IDeleteNodeAction {
 
 export interface IDeleteNodesAction {
   type: `DELETE_NODES`
-  payload: Identifier[]
-  fullNodes: IGatsbyNode[]
+  payload: Array<Identifier>
+  fullNodes: Array<IGatsbyNode>
 }
 
 export interface ISetSiteFlattenedPluginsAction {
@@ -723,14 +725,14 @@ interface IStartIncrementalInferenceAction {
 interface IBuildTypeMetadataAction {
   type: `BUILD_TYPE_METADATA`
   payload: {
-    nodes: IGatsbyNode[]
+    nodes: Array<IGatsbyNode>
     typeName: string
   }
 }
 
 interface IDisableTypeInferenceAction {
   type: `DISABLE_TYPE_INFERENCE`
-  payload: string[]
+  payload: Array<string>
 }
 
 interface ISetProgramAction {
@@ -740,5 +742,5 @@ interface ISetProgramAction {
 
 interface ISetProgramExtensions {
   type: `SET_PROGRAM_EXTENSIONS`
-  payload: string[]
+  payload: Array<string>
 }